ml_model/requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0
